Added Navbar, enhanced movie view & welcome view, added footer
Added a navigation bar that can bring you from the home page, to the movie page, to my linkedin, or log you out.

Added a "featured movie" section to the movie view and made the movie view responsive for smaller devices.

Welcome view only shows you the sign up /login buttons if you are not currently logged in. If you are logged in it displays a message and links to my react version of the project.

Added footer with links to my socials.